# openapi/Skoleoverblikket.Api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Skoleoverblikket API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/classes": {
      "get": {
        "tags": [
          "Classes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassesController_ClassDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassesController_ClassDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassesController_ClassDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Classes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{id}": {
      "get": {
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_UpsertClassRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesController_ClassDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/classes/year-roll": {
      "post": {
        "tags": [
          "Classes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_YearRollRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_YearRollRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClassesController_YearRollRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassesController_ClassDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "gradeLevel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isAccessibleToCurrentUser": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ClassesController_UpsertClassRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "gradeLevel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClassesController_YearRollCreateEntry": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClassesController_YearRollRenameEntry": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "newName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClassesController_YearRollRequest": {
        "type": "object",
        "properties": {
          "renames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassesController_YearRollRenameEntry"
            },
            "nullable": true
          },
          "archive": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "create": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassesController_YearRollCreateEntry"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}